Update and consolidate streaming API
